converted frontend to typescript
