+++ frontend/index.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
+import App from './App';
 import { Provider } from 'react-redux';
-import store from './store/store.js';
+import store from './store/store';
 import { BrowserRouter } from 'react-router-dom';
 
 const container = document.getElementById('root');
-const root = createRoot(container);
+const root = createRoot(container as Element);
 
 root.render(
   <Provider store={store}>
