+++ package.json
@@ -2,7 +2,7 @@
   "name": "rekuberate",
   "version": "1.0.0",
   "description": "kubernetes app",
-  "main": "index.js",
+  "main": "index.tsx",
   "scripts": {
     "frontend": "NODE_ENV=development webpack-dev-server --open --hot --progress --color",
     "dev": "concurrently \"cross-env NODE_ENV=development webpack-dev-server --open --hot --progress --color \" \"cross-env NODE_ENV=development nodemon ./server/server.js\"",
@@ -22,7 +22,7 @@
     "@babel/plugin-transform-runtime": "^7.21.4",
     "@babel/preset-env": "^7.21.5",
     "@babel/preset-react": "^7.18.6",
-    "@babel/runtime": "^7.21.5", 
+    "@babel/runtime": "^7.21.5",
     "babel-loader": "^9.1.2",
     "css-loader": "^6.7.3",
     "html-webpack-plugin": "^5.5.1",
@@ -35,6 +35,10 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
+    "@types/jest": "^29.5.1",
+    "@types/node": "^20.2.5",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
     "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
     "concurrently": "^8.0.1",
@@ -46,6 +50,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^6.11.1"
+    "react-router-dom": "^6.11.1",
+    "ts-loader": "^9.4.3",
+    "typescript": "^5.0.4"
   }
 }
